custom: Refactor and optimize LookupData and CustomTimeUnits, CustomTimeUnit ...
Remove `LookupData::with_multiplier()` and use `new` method to construct this
struct always with a multiplier.

CustomTimeUnit now equals another CustomTimeUnit when the base_unit and the
multiplier are the same. The identifiers don't play a role in the equality check
anymore.

CustomTimeUnits: Unify the `time_units` and `custom_time_units` fields into one
field `time_units`. This simplifies the methods in this struct and in
consequence the `lookup_custom_mut` could be removed too.

Add some missing documentation in the public api.