fundu-core/fundu: Fix: A sign without a number is sometimes parsed as 1 second
This error happened when the configuration options parse_multiple together with
number_is_optional was set. For example "1s + 1s" resulted in a
Duration::positive(3, 0) because the sign was parsed as 1 second instead of
being interpreted as a sign of the next duration.

Such an input results now in an ParseError::InvalidInput