fundu-core: Fix time units parsing to be not final in ReprParserSingle ...
This only affects the parse_time_unit function. parse_number_time_unit still
handles a returned None value from TimeUnitsLike::get as error.

Due to the changes in the parsing of input that is not a whole number or
fraction, it cannot happen anymore that a sign without a number is not an error
in ReprParserTemplate::parse. There is also always something useful returned
by that method and an exponent cannot be without a mantissa.